Deprecate the test_helpers module (this was more of a holdover from older versions of Astropy anyways).  Add multiple ways of getting a ./setup.py test command compatible with different versions of astropy (or lack of astropy).
